Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-pannernode-interface/panner-automation-basic-expected.txt

CONSOLE MESSAGE: line 241: TypeError: undefined is not an object (evaluating 'pannerSetter.value = 1')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'pannerSetter.value = 1')

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [Stereo panner.positionX]  
